requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
Flask-Mail==0.9.1
Flask-Login==0.6.3
python-dotenv==1.0.1
itsdangerous==2.2.0
Werkzeug==3.1.3
requests==2.32.5

